+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 19/02/2021
+### **Added**
+- Add simplest ray tracer shader
+
 ## **[1.4.0]** - 14/10/2020
 ### **Added**
 - Add pixel machine tribute shader